requirements.txt
quart
uvicorn
pycryptodome
PyPDF2
qrcode
pillow
openai
python-dotenv
requests